feat: 删除校验

+++ distribution/service/switch_model.py
@@ -3,7 +3,7 @@ import time
 from django.core.paginator import Paginator
 from django.forms import model_to_dict
 
-from distribution.models import Router, RouterPort, Switch
+from distribution.models import Router, RouterPort, Switch, Workstation
 from ip_distribution.utils.exception_util import BusinessException
 from ip_distribution.utils.log_util import get_logger
 
@@ -108,82 +108,9 @@ class SwitchModel(object):
         return {"count": count, "list": data_list}
 
     def del_switch(self, switch_id):
+        if Workstation.objects.filter(switch_id=switch_id).exists():
+            raise BusinessException("该交换机已被工位使用，不能删除")
         switch = Switch.objects.get(id=switch_id)
         switch.delete()
         logger.info("已删除交换机：{}".format(switch))
-
-
-class SwitchPortModel(object):
-
-    def add(self, router_id, start_addr, end_addr, mask, gateway, dns, username):
-        add_params = {
-            "router_id": router_id,
-            "start_addr": start_addr,
-            "end_addr": end_addr,
-            "mask": mask,
-            "gateway": gateway,
-            "dns": dns,
-            "create_time": int(time.time()),
-            "create_user": username,
-        }
-        logger.info("添加端口信息：{}".format(add_params))
-        RouterPort.objects.create(**add_params)
-        logger.info("添加端口信息：{}".format(username))
-
-    def modity(self, router_port_id, router_id, start_addr, end_addr, mask, gateway, dns):
-        modify_params = {}
-        if router_id:
-            modify_params["router_id"] = router_id
-        if start_addr:
-            modify_params["start_addr"] = start_addr
-        if end_addr:
-            modify_params["end_addr"] = end_addr
-        if mask:
-            modify_params["mask"] = mask
-        if gateway:
-            modify_params["gateway"] = gateway
-        if dns:
-            modify_params["dns"] = dns
-        logger.info("修改路由器端口信息：{}".format(modify_params))
-        RouterPort.objects.filter(id=router_port_id).update(**modify_params)
-
-    def detail(self, port_id):
-        router_port = RouterPort.objects.get(id=port_id)
-        router = Router.objects.get(id=router_port.router_id)
-        router_port_info = model_to_dict(router_port)
-        router_port_info["router_name"] = router.name
-        return router_port_info
-
-    def router_port_list(self, router_id_list, page, size, **kwargs):
-        router_port_list = RouterPort.objects.all().order_by("-id")
-        if router_id_list:
-            router_port_list = router_port_list.filter(router_id__in=router_id_list)
-
-        router_map = {}
-        router_all = Router.objects.all()
-        for item in router_all:
-            router_map[item.id] = item.name
-
-        count = router_port_list.count()
-        paginator = Paginator(router_port_list, size)
-        router_port_list = paginator.get_page(page)
-        data_list = []
-        for item in router_port_list:
-            data_list.append({
-                "router_name": router_map.get(item.router_id),
-                "router_id": item.router_id,
-                "start_addr": item.start_addr,
-                "end_addr": item.end_addr,
-                "mask": item.mask,
-                "gateway": item.gateway,
-                "dns": item.dns,
-                "create_time": item.create_time,
-                "create_user": item.create_user,
-            })
-        return {"count": count, "list": data_list}
-
-    def del_router_port(self, port_id):
-        port = RouterPort.objects.get(id=port_id)
-        port.delete()
-        logger.info("已删除路由器端口：{}".format(port))
